change broken-link-disabler.js to brokenLinkMixin.ts and move to another folder
